requirements.txt
dspy-ai
pandas
python-dotenv
matplotlib
seaborn
rich
datasets
evaluate
scikit-learn
